django_app/requirements/base.txt
Django==2.2.7
django-allauth==0.40.0
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-cors-headers==3.1.0
django-environ==0.4.5
django-modeltranslation==0.13.3
django-rest-auth==0.9.5
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
ipython==7.9.0
postgres==2.2.2
psycopg2-binary==2.8.3
celery==4.3.0
redis==3.3.11